+++ Modelo/src/main/java/com/erpweb/servicios/compras/PedidoService.java
@@ -90,11 +90,13 @@ public class PedidoService {
 	
 	public AccionRespuesta eliminarPedido(Pedido pedido) {
 		
-		logger.debug("Entramos en el metodo eliminarPedido() con ID={}", pedido.getId() );
+		logger.debug("Entramos en el metodo eliminarPedido()" );
 		
 		if(pedido == null || pedido.getId() == null) {
 			
-			return new AccionRespuesta();
+			logger.error("ERROR en el metodo eliminarPedido()");
+			
+			return new AccionRespuesta(-1L, "NOK", Boolean.FALSE);
 		}
 		
 		try {
@@ -102,26 +104,35 @@ public class PedidoService {
 			//Elimnamos el pedido
 			pedidoRepository.deleteById(pedido.getId());
 			
+			return new AccionRespuesta(-2L, "OK", Boolean.TRUE);
+			
 		}catch(Exception e) {
 			
 			logger.error("Error en el metodo eliminarPedido() con ID={} ", pedido.getId() );
 			
 			e.printStackTrace();
 						
-			return new AccionRespuesta();
+			return new AccionRespuesta(-1L, "NOK", Boolean.FALSE);
 		}
-				
-		return new AccionRespuesta();
 	}
 	
 	public AccionRespuesta eliminarPedidoPorId(Long pedidoId) {
 		
-		logger.error("Entramos en el metodo eliminarPedidoPorId() con id={}", pedidoId );
+		logger.error("Entramos en el metodo eliminarPedidoPorId()" );
+		
+		if( pedidoId == null ) {
+			
+			logger.error("ERROR en el metodo eliminarPedidoPorId()");
+			
+			return new AccionRespuesta(-1L, "NOK", Boolean.FALSE);
+		}
 				
 		try {
 			
 			//Elimnamos el pedido
 			pedidoRepository.deleteById(pedidoId);
+			
+			return new AccionRespuesta(-2L, "OK", Boolean.TRUE);
 			
 		}catch(Exception e) {
 			
@@ -129,10 +140,8 @@ public class PedidoService {
 			
 			e.printStackTrace();
 			
-			return new AccionRespuesta();
+			return new AccionRespuesta(-1L, "NOK", Boolean.FALSE);
 		}
-		
-		return new AccionRespuesta();
 	}
 		
 	public PedidoDto obtenerPedidoDtoDesdePedido(Long id) {
